# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinSight API",
    "description": "IPO X-Ray and verified chat. Contract: docs/06_API_CONTRACT.md.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/ipos": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "List Ipos",
        "description": "The showcase IPO library, filtered and sorted.",
        "operationId": "list_ipos_api_ipos_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Company or sector search.",
              "title": "Q"
            },
            "description": "Company or sector search."
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "listing_date",
                "issue_size"
              ],
              "type": "string",
              "default": "listing_date",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IpoSummary"
                  },
                  "title": "Response List Ipos Api Ipos Get"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/ipos/{id}": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "Get Ipo",
        "description": "One showcase IPO's documents and sections.",
        "operationId": "get_ipo_api_ipos__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpoDetail"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/ipos/{id}/xray": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "Get Xray",
        "description": "The X-Ray fact sheet of one showcase IPO.",
        "operationId": "get_xray_api_ipos__id__xray_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XRayResponse"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/ipos/{id}/pages/{n}": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "Get Page Image",
        "description": "A rendered page as WebP (``w`` for a thumbnail); cached for a year.",
        "operationId": "get_page_image_api_ipos__id__pages__n__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "N"
            }
          },
          {
            "name": "doc",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "rhp",
                "drhp",
                "prospectus"
              ],
              "type": "string",
              "description": "Which document: the RHP or the final Prospectus.",
              "default": "rhp",
              "title": "Doc"
            },
            "description": "Which document: the RHP or the final Prospectus."
          },
          {
            "name": "w",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1200,
                  "minimum": 40
                },
                {
                  "type": "null"
                }
              ],
              "description": "Width in pixels (thumbnail)",
              "title": "W"
            },
            "description": "Width in pixels (thumbnail)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/ipos/{id}/pages/{n}/words": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "Get Page Words",
        "description": "The words and boxes of one page.",
        "operationId": "get_page_words_api_ipos__id__pages__n__words_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "N"
            }
          },
          {
            "name": "doc",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "rhp",
                "drhp",
                "prospectus"
              ],
              "type": "string",
              "description": "Which document: the RHP or the final Prospectus.",
              "default": "rhp",
              "title": "Doc"
            },
            "description": "Which document: the RHP or the final Prospectus."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageWords"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/ipos/{id}/suggested-questions": {
      "get": {
        "tags": [
          "ipos"
        ],
        "summary": "Suggested Questions",
        "description": "The question chips for one IPO.",
        "operationId": "suggested_questions_api_ipos__id__suggested_questions_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestedQuestion"
                  },
                  "title": "Response Suggested Questions Api Ipos  Id  Suggested Questions Get"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "description": "Answer a question as server-sent events, replaying the demo cache in demo mode.",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-sent events; see the x-sse-events map for each payload model.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "x-sse-events": {
                  "stage": {
                    "$ref": "#/components/schemas/StageEvent"
                  },
                  "guard": {
                    "$ref": "#/components/schemas/GuardEvent"
                  },
                  "retrieval": {
                    "$ref": "#/components/schemas/RetrievalEvent"
                  },
                  "abstain": {
                    "$ref": "#/components/schemas/AbstainEvent"
                  },
                  "token": {
                    "$ref": "#/components/schemas/TokenEvent"
                  },
                  "answer": {
                    "$ref": "#/components/schemas/AnswerEvent"
                  },
                  "verdict": {
                    "$ref": "#/components/schemas/VerdictEvent"
                  },
                  "final": {
                    "$ref": "#/components/schemas/FinalEvent"
                  },
                  "error": {
                    "$ref": "#/components/schemas/ErrorEvent"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/voice": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Voice",
        "description": "Transcribe a short spoken question (at most 20 s).",
        "operationId": "voice_api_voice_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_voice_api_voice_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/traces/{trace_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Trace",
        "description": "The full trace of one chat answer.",
        "operationId": "get_trace_api_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trace"
                }
              }
            }
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        }
      }
    },
    "/api/lab/ladder": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Ladder",
        "description": "Model Lab: the extractor ladder results.",
        "operationId": "lab_ladder_api_lab_ladder_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Ladder Api Lab Ladder Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/fields": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Fields",
        "description": "Model Lab: per-field X-Ray accuracy.",
        "operationId": "lab_fields_api_lab_fields_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Fields Api Lab Fields Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/verifier": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Verifier",
        "description": "Model Lab: the number verifier's seeded-error results.",
        "operationId": "lab_verifier_api_lab_verifier_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Verifier Api Lab Verifier Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/weaklabels": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Weaklabels",
        "description": "Model Lab: weak-label statistics and the audit precision.",
        "operationId": "lab_weaklabels_api_lab_weaklabels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Weaklabels Api Lab Weaklabels Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/frontier": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Frontier",
        "description": "Model Lab: FinSight against frontier chatbots on the same questions (E9).",
        "operationId": "lab_frontier_api_lab_frontier_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Frontier Api Lab Frontier Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/retrieval": {
      "get": {
        "tags": [
          "lab"
        ],
        "summary": "Lab Retrieval",
        "description": "Model Lab: retrieval recall by method.",
        "operationId": "lab_retrieval_api_lab_retrieval_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Lab Retrieval Api Lab Retrieval Get"
                }
              }
            }
          },
          "4XX": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IpoSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "listing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Date"
          },
          "rhp_pages": {
            "type": "integer",
            "title": "Rhp Pages"
          },
          "issue_size_inr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Size Inr"
          },
          "fresh_inr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fresh Inr"
          },
          "ofs_inr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ofs Inr"
          },
          "xray_status": {
            "type": "string",
            "enum": [
              "ready",
              "building",
              "missing"
            ],
            "title": "Xray Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company",
          "rhp_pages",
          "xray_status"
        ],
        "title": "IpoSummary",
        "description": "A showcase IPO in the library list, with its headline sizes."
      },
...
    }
  }
}